+++ app/clients/mongodb.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from motor.motor_asyncio import AsyncIOMotorClient
 
 MONGO_URL = "mongodb://localhost:27017"
-MONGO_URL_EC2 = "mongodb://54.180.4.224:27017"
+#MONGO_URL_EC2 = "mongodb://54.180.4.224:27017"
 client = AsyncIOMotorClient(MONGO_URL)
 db = client["ai_platform"]
