+++ app/routers/post_assessment_router.py
@@ -6,8 +6,9 @@ from typing import List
 
 from app.models.pre_assessment.request import AssessmentResult
 from app.models.pre_assessment.response import QuestionStructure
-from app.services.assessment.common import get_user, subject_id_to_name, result_generate, safe_sample
+from app.services.assessment.common import get_user, result_generate, safe_sample
 from app.services.assessment.post import generate_key
+from app.services.common.common import subject_id_to_name
 
 router = APIRouter()
