+++ app/routers/pre_assessment_router.py
@@ -4,7 +4,8 @@ from datetime import datetime
 
 from app.models.pre_assessment.request import AssessmentInput, AssessmentResult
 from app.models.pre_assessment.response import QuestionStructure
-from app.services.assessment.common import get_user, subject_id_to_name, safe_sample, result_generate
+from app.services.assessment.common import get_user, safe_sample, result_generate
+from app.services.common.common import subject_id_to_name
 from app.utils.level_utils import calculate_level_from_answers
 from typing import List
